add 3 files and move some functions to the correct files (and aligned comments)
