+++ apis/fluent/atrium-api-fluent/src/commonMain/kotlin/ch/tutteli/atrium/api/fluent/en_GB/iteratorExpectations.kt
@@ -1,7 +1,9 @@
 package ch.tutteli.atrium.api.fluent.en_GB
 
 import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.logic.*
+import ch.tutteli.atrium.logic._logicAppend
+import ch.tutteli.atrium.logic.hasNext
+import ch.tutteli.atrium.logic.hasNotNext
 
 /**
  * Expects that the subject of `this` expectation (an [Iterator]) has a next element.
@@ -26,29 +28,3 @@ fun <T : Iterator<*>> Expect<T>.toHaveNext(): Expect<T> =
  */
 fun <T : Iterator<*>> Expect<T>.notToHaveNext(): Expect<T> =
     _logicAppend { hasNotNext() }
-
-/**
- * Expects that `this` expectation (an [Iterator]) has a next element and
- * returns an [Expect] for the next element.
- *
- * @return The newly created [Expect] for the next element.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.IteratorExpectationSamples.nextFeature
- *
- * @since 1.3.0
- */
-fun <E, T : Iterator<E>> Expect<T>.next(): Expect<E> =
-    _logic.next().transform()
-
-/**
- * Expects that `this` expectation (an [Iterator]) has a next element and that
- * the next element holds all assertions the given [assertionCreator] creates for it.
- *
- * @return an [Expect] for the subject of the next expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.IteratorExpectationSamples.next
- *
- * @since 1.3.0
- */
-fun <E, T : Iterator<E>> Expect<T>.next(assertionCreator: Expect<E>.() -> Unit): Expect<T> =
-    _logic.next().collectAndAppend(assertionCreator)
