+++ apis/infix/atrium-api-infix/src/commonMain/kotlin/ch/tutteli/atrium/api/infix/en_GB/iteratorExpectations.kt
@@ -1,7 +1,9 @@
 package ch.tutteli.atrium.api.infix.en_GB
 
 import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.logic.*
+import ch.tutteli.atrium.logic._logicAppend
+import ch.tutteli.atrium.logic.hasNext
+import ch.tutteli.atrium.logic.hasNotNext
 
 /**
  * Expects that the subject of `this` expectation (an [Iterator]) has a next element.
@@ -26,29 +28,3 @@ infix fun <T : Iterator<*>> Expect<T>.toHave(@Suppress("UNUSED_PARAMETER") next:
  */
 infix fun <T : Iterator<*>> Expect<T>.notToHave(@Suppress("UNUSED_PARAMETER") next: next): Expect<T> =
     _logicAppend { hasNotNext() }
-
-/**
- * Expects that `this` expectation (an [Iterator]) has a next element and
- * returns an [Expect] for the next element.
- *
- * @return The newly created [Expect] for the next element.
- *
- * @sample TODO
- *
- * @since 1.3.0
- */
-infix fun <E, T : Iterator<E>> Expect<T>.next(@Suppress("UNUSED_PARAMETER") next: next): Expect<E> =
-    _logic.next().transform()
-
-/**
- * Expects that `this` expectation (an [Iterator]) has a next element and that
- * the next element holds all assertions the given [assertionCreator] creates for it.
- *
- * @return an [Expect] for the subject of the next expectation.
- *
- * @sample TODO
- *
- * @since 1.3.0
- */
-infix fun <E, T : Iterator<E>> Expect<T>.next(assertionCreator: Expect<E>.() -> Unit): Expect<T> =
-    _logic.next().collectAndAppend(assertionCreator)
